+++ alicompat/fabric/src/compat/ribbits/java/com/yanny/alicompat/compat/ribbits/RibbitsLang.java
@@ -1,6 +1,7 @@
 package com.yanny.alicompat.compat.ribbits;
 
 import com.yanny.alicompat.ICompatTranslations;
+import com.yanny.alicompat.Utils;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;