+++ alicompat/neoforge/src/compat/grimoireofgaia/java/com/yanny/alicompat/compat/grimoireofgaia/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package com.yanny.alicompat.compat.grimoireofgaia;
+
+import javax.annotation.ParametersAreNonnullByDefault;