+++ alicompat/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added Repurposed Structures support
 - Added Sophisticated Backpacks support
 - Added Sophisticated Storage support
+- Translation keys now carry target mod id
 
 ## [1.0.0]